+++ src/libstd/sys/redox/mod.rs
@@ -17,6 +17,7 @@ pub use self::rand::hashmap_random_keys;
 pub mod args;
 #[cfg(feature = "backtrace")]
 pub mod backtrace;
+pub mod cmath;
 pub mod condvar;
 pub mod env;
 pub mod ext;
